% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmgm-package.R
\docType{package}
\name{mmgm-package}
\alias{mmgm}
\alias{mmgm-package}
\title{mmgm: Estimates the degree of decomposition of peat with a mixing model for \eqn{\gamma_\text{MIRS}}}
\description{
Provides a mixing model for \eqn{\gamma_\text{MIRS}} bulk estimates (the degree of decomposition, \eqn{\gamma} estimated from mid infrared spectra) of peat samples. The mixing model takes \eqn{\gamma_\text{MIRS}} and the mass fractions of each component of the sample as input and estimates \eqn{\gamma} of each component and of the entire sample. Batch processing is possible.
}
\author{
\strong{Maintainer}: Henning Teickner \email{henning.teickner@uni-muenster.de}

}
\keyword{internal}
