% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmgm_b_Intercept}
\alias{mmgm_b_Intercept}
\title{Intercepts from the \code{irpeatmodels} prediction models for \eqn{\gamma_\text{MIRS}}}
\format{
\subsection{\code{mmgm_b_Intercept}}{

A data frame with 3 rows and 4 columns:
\describe{
\item{\code{model_id}}{Integer value. The identifier of the model in the
\code{irpeatmodels} package.}
\item{\verb{model name}}{Character value. The name of the model in the
\code{irpeatmodels} package.}
\item{\code{b_intercept_p1}}{Numeric value. The average of the intercept of the
model.}
\item{\code{b_intercept_p2}}{Numeric value. The standard deviation of the
intercept of the model.}
}
}
}
\usage{
mmgm_b_Intercept
}
\description{
This dataset is used internally in the mixing model.
}
\keyword{datasets}
