% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmgm_estimate_mixing_model.R
\name{mmgm_estimate_gamma_mirs_mixing_1}
\alias{mmgm_estimate_gamma_mirs_mixing_1}
\title{Estimates the mixing model from \insertCite{Teickner.2025h;textual}{mmgm}}
\usage{
mmgm_estimate_gamma_mirs_mixing_1(stan_data, ...)
}
\arguments{
\item{stan_data}{An object as returned by \code{mmgm_make_stan_data()}.}

\item{...}{Named arguments to the \code{sample()} method of CmdStan model
objects: \url{https://mc-stan.org/cmdstanr/reference/model-method-sample.html}}
}
\value{
A \code{CmdStanMCMC} object with the estimated mixing model. The model has
the following parameters:
\describe{
\item{\code{gamma_component}}{The degree of decomposition (\eqn{\gamma}) of each
component (the fraction of the initial mass remaining in g g\eqn{^{-1}}).}
\item{\code{gamma_mixture}}{\eqn{\gamma} of each mixture.}
\item{\code{gamma_mirs_mixture_obs}}{Average \eqn{\gamma_\text{MIRS}} of the
mixtures.}
\item{\code{phi}}{A scaled version of the scale parameter of the beta
distribution for \eqn{\gamma_\text{MIRS}} of the mixtures.}
\item{\code{b_intercept}}{Intercept of the linear model that predicts
\eqn{\gamma_\text{MIRS}} from mid-infrared spectra.}
}
}
\description{
Estimates the mixing model from \insertCite{Teickner.2025h;textual}{mmgm}
}
\examples{
if (instantiate::stan_cmdstan_exists()) {

cmdstanr::check_cmdstan_toolchain(fix = FALSE, quiet = TRUE)

stan_data <-
  mmgm_make_stan_data(
    x = mmgm_example_data,
    id_model_gamma_mirs = 2L,
    priors = mmgm_make_default_priors(x = mmgm_example_data)
  )

fit <-
  mmgm_estimate_gamma_mirs_mixing_1(
    stan_data,
    iter_sampling = 50,
    iter_warmup = 30,
    chains = 2,
    sig_figs = 14,
    seed = 7667
  )

 fit$summary()
}

}
\references{
\insertAllCited{}
}
\concept{mmgm_estimate_mixing_model}
