% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmgm_example_data}
\alias{mmgm_example_data}
\title{Example data in the \code{mmgm} package}
\format{
\subsection{\code{mmgm_example_data}}{

A list with two elements. The list has the same structure as argument \code{x} in
\code{mmgm_make_stan_data()}.
}
}
\usage{
mmgm_example_data
}
\description{
Example data in the \code{mmgm} package
}
\keyword{datasets}
