% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmgm_make_default_priors.R
\name{mmgm_make_default_priors}
\alias{mmgm_make_default_priors}
\title{Creates a list with default parameter values for a given set of mixtures and components}
\usage{
mmgm_make_default_priors(x)
}
\arguments{
\item{x}{A list with the same structure as argument \code{x} in
\code{mmgm_make_stan_data()}.}
}
\description{
Creates a list with default parameter values for a given set of mixtures and components
}
