% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmgm_make_stan_data.R
\name{mmgm_make_stan_data}
\alias{mmgm_make_stan_data}
\title{Prepares the list to use as data input to the mixing model}
\usage{
mmgm_make_stan_data(
  x,
  id_model_gamma_mirs,
  priors = mmgm_make_default_priors(x = x)
)
}
\arguments{
\item{x}{A list with two elements:
\describe{
\item{\code{components}}{A data frame with a row for each component in the
mixtures and the following columns:
\describe{
\item{\code{id_mixture}}{Unique integer for each mixture.}
\item{\code{id_component}}{Unique integer for each component within a mixture
(the first component of each mixture has \code{id_component == 1}).}
\item{\code{w}}{The mass fraction of the component in the mixture. \code{w} must
sum to 1 across each value of \code{id_mixture}.}
}}
\item{\code{mixtures}}{A data frame with a row for each mixture and the
following columns:
\describe{
\item{\code{id_mixture}}{Unique integer for each mixture.}
\item{\code{degree_of_decomposition_1}, \code{degree_of_decomposition_2}, or
\code{degree_of_decomposition_3}}{Predictions of \eqn{\gamma_\text{MIRS}} for
each mixture, as created with \code{irpeat::irp_degree_of_decomposition_1()},
\code{irpeat::irp_degree_of_decomposition_2()} or
\code{irpeat::irp_degree_of_decomposition_3()}. One or more of the columns may
be present. The columns must be \code{quantities} objects. To create such
objects, use argument \code{do_summary = TRUE} in \code{irpeat::irp_degree_of_decomposition_1()}
or one of the other predicition functions.}
}}
}}

\item{id_model_gamma_mirs}{An integer specifying which \eqn{\gamma_\text{MIRS}}
predictions to use in the mixing model. The corresponding column must be
present in \code{x}. For example if \code{id_model_gamma_mirs == 1}, then values of
column \code{degree_of_decomposition_1} in \code{x} will be used to estimate the mixing
model.}

\item{priors}{A list specifying prior parameters for the mixing model. See
\code{mmgm_default_priors} for the structure of this list.}
}
\value{
A list that can be used as input to \code{mmgm_estimate_mixing_model()}.
}
\description{
Prepares the list to use as data input to the mixing model
}
\examples{
mmgm_make_stan_data(
  x = mmgm_example_data,
  id_model_gamma_mirs = 2L,
  priors = mmgm_make_default_priors(x = mmgm_example_data)
)

}
